autoform-ai-configurator/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
pillow
python-dotenv
